requirements.txt
fastapi~=0.101.1
pydantic~=2.2.1
celery~=5.3.1
certifi~=2023.7.22
pymongo~=4.4.1
redis~=5.0.0
python-dateutil~=2.8.2
requests~=2.31.0
pytz~=2023.3
pika~=0.13.0